# Update SoundFile.blocks.