--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,6 @@ dist/
 build/
 PySoundFile.egg-info/
 .eggs/
+*.egg-info/
+.cache/
+.vscode/