--- a/tests/test_pysoundfile.py
+++ b/tests/test_pysoundfile.py
@@ -371,7 +371,7 @@ def test_blocks_with_frames_and_fill_value(file_stereo_r):
 def test_blocks_with_out(file_stereo_r):
     out = np.empty((3, 2))
     blocks = list(sf.blocks(file_stereo_r, out=out))
-    assert blocks[0] is out
+    assert blocks[0].base is out
     # First frame was overwritten by second block:
     assert np.all(blocks[0] == data_stereo[[3, 1, 2]])
 
@@ -1006,10 +1006,6 @@ def test_write_non_seekable_file(file_w):
         with pytest.raises(ValueError) as excinfo:
             f.read()
         assert "frames" in str(excinfo.value)
-
-        with pytest.raises(ValueError) as excinfo:
-            list(f.blocks(blocksize=3, overlap=1))
-        assert "overlap" in str(excinfo.value)
 
     data, fs = sf.read(filename_new, dtype='int16')
     assert np.all(data == data_mono)