This PR removes the necessity to seek in the input file when generating blocks because seeking is relatively expensive. Here is a benchmark on an [ACDC snippet](https://en.wikipedia.org/wiki/File:ACDC_-_Back_In_Black-sample.ogg) (25 seconds) and a [metal podcast](http://archive.org/download/OpenMetalcastEpisode162InfinitePossibilities/open_metalcast_162.ogg) (32 mins, 15 seconds).

```python
# profile.py
import soundfile
import tqdm
import sys

with soundfile.SoundFile(sys.argv[1]) as sf:
    for _ in tqdm.tqdm(sf.blocks(6 * sf.samplerate, sf.samplerate)):
        pass
```

```
$ python profile.py ACDC_-_Back_In_Black-sample.ogg
5it [00:00, 20.51it/s]  # on master
$ python profile.py ACDC_-_Back_In_Black-sample.ogg
5it [00:00, 30.96it/s]  # with this patch
```

```
$ python profile.py open_metalcast_162.ogg
214it [05:03,  2.69s/it]  # on master (cancelled after five minutes)
$ python profile.py open_metalcast_162.ogg
5.379634857177734  # with this patch
```

One interesting observation: The code on `master` starts of a bit slower than the patched version on long files but becomes much slower as it processes more of the file. Maybe because the implementation seeks from the beginning of the file rather than doing a relative seek from the current position?