@mgeier 2017-10-04
I'm currently on the road and can't test it myself but doesn't that
pessimize the IMHO more important case where overlap is zero (and out is
None)?

And the current call to seek() doesn't seek from the beginning but
uses SEEK_CUR. I don't know what's going wrong when progressing further
into a large file. Probably a memory leak of some sort?
Did you notice increased memory usage?
I guess this phenomenon would be worth investigating further ...

Which OS did you test on?

@bastibe Did you get the same performance results?