@tillahoffmann 2017-10-04
@mgeier, I don't see a performance degradation for zero overlap on my end (with an updated profiling script):

```python
# pysound_profile.py
import argparse
import hashlib
import platform
import soundfile
import tqdm

ap = argparse.ArgumentParser("pysound_profile")
ap.add_argument('filenames', nargs='+')
ap.add_argument('--overlap', '-o', type=float, default=1)
ap.add_argument('--block-size', '-b', type=float, default=6)
ap.add_argument('--hash', '-d', default='sha256')
args = ap.parse_args()

print("pysound_profile")
print(platform.uname())
print()

for filename in args.filenames:
    print("Processing '%s' with block size %fs and overlap %fs..." %
          (filename, args.block_size, args.overlap))
    with soundfile.SoundFile(filename) as sf:
        hasher = hashlib.new(args.hash)
        blocks = sf.blocks(
            round(args.block_size * sf.samplerate),
            round(args.overlap * sf.samplerate)
        )
        for block in tqdm.tqdm(blocks):
            hasher.update(block)

    print("%s = %s" % (args.hash, hasher.digest().hex()))
```

```
(master) $ python pysound_profile.py -o 0 ACDC_-_Back_In_Black-sample.ogg open_metalcast_162.ogg
uname_result(system='Darwin', node='Tills-MacBook-Pro-2.local', release='17.0.0', version='Darwin Kernel Version 17.0.0: Thu Aug 24 21:48:19 PDT 2017; root:xnu-4570.1.46~2/RELEASE_X86_64', machine='x86_64', processor='i386')

Processing 'ACDC_-_Back_In_Black-sample.ogg' with block size 6.000000s and overlap 0.000000s...
5it [00:00, 27.42it/s]
sha256 = 4737d4ebc248822430df008d99cc1163f8800b41c237101163aebaa0fb370d15
Processing './open_metalcast_162.ogg' with block size 6.000000s and overlap 0.000000s...
323it [00:08, 38.26it/s]
sha256 = 7e8d690a6b580379c083b3a4817db74c348c84ee42b093ffe572cf72d4ac39f0

(blocks) $ pysound_profile
uname_result(system='Darwin', node='Tills-MacBook-Pro-2.local', release='17.0.0', version='Darwin Kernel Version 17.0.0: Thu Aug 24 21:48:19 PDT 2017; root:xnu-4570.1.46~2/RELEASE_X86_64', machine='x86_64', processor='i386')

Processing 'ACDC_-_Back_In_Black-sample.ogg' with block size 6.000000s and overlap 0.000000s...
5it [00:00, 24.51it/s]
sha256 = 4737d4ebc248822430df008d99cc1163f8800b41c237101163aebaa0fb370d15
Processing './open_metalcast_162.ogg' with block size 6.000000s and overlap 0.000000s...
323it [00:08, 39.24it/s]
sha256 = 7e8d690a6b580379c083b3a4817db74c348c84ee42b093ffe572cf72d4ac39f0
```

I don't see a memory leak on my end rerunning the blocks on master with overlap.