@bastibe 2017-10-05
@mgeier A priori, I would imagine that seeking is approximately free for uncompressed files, but expensive for compressed files. Regardless, I would think that shuffling some numpy memory should always be a bit less expensive than calling out to a C library and updating file pointers.

Let's see how my intuition compares to reality ;-)

Current Master:

```python
>>> %timeit for block in soundfile.blocks('still_alive.wav', 512, 0): pass
773 ms ± 3.53 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

>>> %timeit for block in soundfile.blocks('still_alive.wav', 512, 256): pass
1.58 s ± 13.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

>>> %timeit for block in soundfile.blocks('still_alive.flac', 512, 0): pass
9.51 s ± 83 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

>>> %timeit for block in soundfile.blocks('still_alive.flac', 512, 256): pass
36.1 s ± 422 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Your branch:

```python
>>> %timeit for block in soundfile.blocks('still_alive.wav', 512, 0): pass
622 ms ± 22.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

>>> %timeit for block in soundfile.blocks('still_alive.wav', 512, 256): pass
1.21 s ± 4.69 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

>>> %timeit for block in soundfile.blocks('still_alive.flac', 512, 0): pass
9.13 s ± 128 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

>>> %timeit for block in soundfile.blocks('still_alive.flac', 512, 256): pass
18.3 s ± 47.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

As you can see, there is a modest performance benefit for non-compressed files or zero overlap (5%-30%), but a substantial (2x) benefit for compressed files with nonzero overlap! 

This performance benefit, and getting rid of the `seek` for nonseekable files, makes this very worthwhile in my eyes. Do you have any objections to merging this, @mgeier?

- - -

As a side note, I would not have thought that block processing takes so much time! Simply reading the whole file takes

```python
>>> %timeit soundfile.read('still_alive.wav');
70.2 ms ± 258 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

>>> %timeit soundfile.read('still_alive.flac');
235 ms ± 3.97 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

while reading a single block takes

```python
>>> %timeit soundfile.read('still_alive.wav', start=44100, frames=512);
160 µs ± 230 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> %timeit soundfile.read('still_alive.flac', start=44100, frames=512);
811 µs ± 2.64 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

which this adds up to the 18-36 s of processing time if you multiply it with 30k blocks. 

It might be worthwhile for `blocks` to read larger chunks and split them into blocks internally. Is anyone interested in trying this?